rename network-helper -> networkHelper
