+++ src/electron/helper/networkHelper.js
